add 64-bit limit API for larger tick budgets

+++ src/internal.hpp
@@ -1511,16 +1511,16 @@ struct Internal {
   // Internal functions to enable explicit search limits.
   //
   void limit_terminate (int);
-  void limit_decisions (int);     // Force decision limit.
-  void limit_conflicts (int);     // Force conflict limit.
-  void limit_preprocessing (int); // Enable 'n' preprocessing rounds.
-  void limit_local_search (int);  // Enable 'n' local search rounds.
+  void limit_decisions (int64_t);     // Force decision limit.
+  void limit_conflicts (int64_t);     // Force conflict limit.
+  void limit_preprocessing (int64_t); // Enable 'n' preprocessing rounds.
+  void limit_local_search (int64_t);  // Enable 'n' local search rounds.
   void limit_ticks (int64_t);     // Force ticks limit.
 
   // External versions can access limits by 'name'.
   //
   static bool is_valid_limit (const char *name);
-  bool limit (const char *name, int); // 'true' if 'name' valid
+  bool limit (const char *name, int64_t); // 'true' if 'name' valid
 
   // Set all the CDCL search limits and increments for scheduling
   // inprocessing, restarts, clause database reductions, etc.
@@ -1551,10 +1551,10 @@ struct Internal {
   //
   int already_solved ();
   int restore_clauses ();
-  bool preprocess_round (int round);
+  bool preprocess_round (int64_t round);
   void preprocess_quickly (bool always);
   int preprocess (bool always);
-  int local_search_round (int round);
+  int local_search_round (int64_t round);
   int local_search ();
   int lucky_phases ();
   int cdcl_loop_with_inprocessing ();
